requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0
tld==0.13
trafilatura==2.0.0
courlan==1.3.2
pymupdf>=1.23.7
beautifulsoup4==4.13.3
requests>=2.31.0
pdfminer.six==20221105
tzlocal==5.3
anthropic==0.3.11
google-generativeai>=0.3.1
PyPDF2>=3.0.0
python-docx>=1.0.1
fpdf>=1.7.2
pdfplumber>=0.10.3
matplotlib>=3.8.2
seaborn>=0.13.0
Pillow>=10.1.0
python-dotenv>=1.0.0
scikit-learn>=1.3.2
nltk>=3.8.1

# Compliance Module Dependencies
python-dateutil>=2.8.2
dataclasses-json>=0.5.7
regex>=2023.0.0

plotly>=5.18.0

# RAG and ML Dependencies
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.2
transformers>=4.35.0
datasets>=2.14.0
faiss-cpu>=1.7.4
pypdf>=3.15.1
huggingface_hub>=0.19.0
torch>=2.0.0
chromadb>=0.4.18
